move app and main
