+++ src/app/App.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Chat from "./lib/components/chat/Chat.svelte";
+  import Chat from "./components/chat/Chat.svelte";
 </script>
 
 <main>
