+++ src/app/main.ts
@@ -1,5 +1,5 @@
 import App from "./App.svelte";
-import "./global.css";
+import "../global.css";
 
 const app = new App({
   // target content as it always exists on the youtube ui
